+++ src/application.rs
@@ -1,6 +1,7 @@
 use crate::default_manifest::DEFAULT_MANIFEST_FILE;
 use crate::project::{Project, ProjectAction, ProjectFileAction};
-use log::{debug, error};
+use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
+use std::collections::VecDeque;
 use std::fs::File;
 use std::io::Write;
 use std::path::{Path, PathBuf};
@@ -8,7 +9,7 @@ use std::process::Command;
 use std::sync::mpsc::channel;
 use std::sync::mpsc::Sender;
 use std::sync::{Arc, Mutex};
-use std::thread;
+use std::{fs, thread};
 
 #[cfg(target_os = "linux")]
 use std::os::unix::fs::symlink;
@@ -29,31 +30,9 @@ pub const SYNC: &str = "sync";
 pub const PIN: &str = "pin";
 
 pub const LIGHT: &str = "light";
+pub const QUIET: &str = "quiet";
+pub const FORCE: &str = "force";
 pub const HTTPS: &str = "https";
-
-#[derive(Debug, Clone)]
-#[repr(u8)]
-pub enum ExitCode {
-    Success = 0,
-    NoManifest = 100,
-    GenerationFailed = 101,
-    ManifestInvalid = 102,
-    CloneFailed = 103,
-    CheckoutFailed = 104,
-    GetCommitIdFailed = 105,
-    WorkdirFailed = 106,
-    CreateSymlinkFailed = 107,
-    CopyFailed = 108,
-    RemoveFailed = 109,
-    SaveFailed = 110,
-    DependencyMissing = 111,
-}
-
-impl From<ExitCode> for std::process::ExitCode {
-    fn from(code: ExitCode) -> Self {
-        std::process::ExitCode::from(code as u8)
-    }
-}
 
 pub enum DwlMode {
     HTTPS,
@@ -62,32 +41,38 @@ pub enum DwlMode {
 
 pub trait ManifestParser {
     /// Parse a manifest file and return a vector of projects.
-    fn parse(&self, file: &str) -> Result<Vec<Project>, ExitCode>;
+    fn parse(&self, file: &str) -> Result<Vec<Project>, String>;
 
     /// Compose a manifest file from a vector of projects.
-    fn compose(&self, projects: &[Project]) -> Result<String, ExitCode>;
+    fn compose(&self, projects: &[Project]) -> Result<String, String>;
 }
 
 /// VersionControl is a trait that defines the methods to interact with a version control system.
 /// The manifest file path is used to avoid changing the working directory which may results in
 /// unexpected behavior when multiple threads are used.
+/// When using the lightweight option, the revision MUST be a branch or a tag. Commit ID are not supported.
 pub trait VersionControl {
     /// Clone a project from a URI to a path.
-    fn clone(&self, manifest_dir: &str, project: &Project, mode: &DwlMode) -> Result<(), ExitCode>;
-
-    /// Clone a project from a URI to a path with a specific revision. The revision MUST be a branch or a tag. Commit ID are not supported.
-    fn clone_lightweight(
+    fn clone(
         &self,
         manifest_dir: &str,
         project: &Project,
         mode: &DwlMode,
-    ) -> Result<(), ExitCode>;
+        pb: Option<&ProgressBar>,
+        lightweight: bool,
+    ) -> Result<(), String>;
 
     /// Update a project to a specific revision.
-    fn checkout(&self, manifest_dir: &str, project: &Project) -> Result<(), ExitCode>;
+    fn checkout(
+        &self,
+        manifest_dir: &str,
+        project: &Project,
+        pb: Option<&ProgressBar>,
+        force: bool,
+    ) -> Result<(), String>;
 
     /// Status of a project.
-    fn get_commit_id(&self, manifest_dir: &str, project: &Project) -> Result<String, ExitCode>;
+    fn get_commit_id(&self, manifest_dir: &str, project: &Project) -> Result<String, String>;
 }
 
 pub struct ManifestInstance {
@@ -97,7 +82,7 @@ pub struct ManifestInstance {
 }
 
 impl ManifestInstance {
-    pub fn new(filename: Option<&String>) -> Result<Self, ExitCode> {
+    pub fn new(filename: Option<&String>) -> Result<Self, String> {
         let (name, file) = read_manifest(filename)?;
 
         Ok(Self {
@@ -119,32 +104,65 @@ impl ManifestInstance {
         &self.projects
     }
 
-    pub fn parse(&mut self, parser: &dyn ManifestParser) -> Result<(), ExitCode> {
+    pub fn parse(&mut self, parser: &dyn ManifestParser) -> Result<(), String> {
         self.projects = parser.parse(&self.file)?;
         Ok(())
     }
 
-    pub fn sync<T>(&self, vcs: &T, mode: &DwlMode) -> Result<(), ExitCode>
+    pub fn sync<T>(
+        &self,
+        vcs: &T,
+        mode: &DwlMode,
+        lightweight: bool,
+        quiet: bool,
+        force: bool,
+    ) -> Result<(), String>
     where
         T: VersionControl + Sync,
     {
+        // Prepare progress bar
+        let multi_progress = MultiProgress::new();
+        let style = ProgressStyle::default_bar()
+            .template("{spinner:.white} [{elapsed_precise}] {bar:40} {pos:>3}/{len:3} {msg}")
+            .expect("Failed to create progress bar style")
+            .progress_chars("##-");
+
+        let mut progress_bars: Vec<Option<ProgressBar>> = Vec::new();
+
+        for _ in self.projects.iter() {
+            let pb = if !quiet {
+                let instance = multi_progress.add(ProgressBar::new(100));
+                instance.set_style(style.clone());
+                Some(instance)
+            } else {
+                None
+            };
+
+            progress_bars.push(pb);
+        }
+
+        // Create channel to save result
         let (tx, rx) = channel();
         let tx = Arc::new(Mutex::new(tx));
 
+        // Spawn a thread for each project
         thread::scope(|s| {
             for project in self.projects.iter() {
                 let tx = Arc::clone(&tx);
 
+                let pb = progress_bars.pop().expect("Failed to get progress bar");
+
                 s.spawn(move || {
                     let dir = self.get_manifest_dir();
                     let mut result = Ok(());
+
                     if is_ok_to_clone(&dir, project.get_path()) {
-                        result = vcs.clone(&dir, project, mode);
+                        result = vcs.clone(&dir, project, mode, pb.as_ref(), lightweight);
                         send_result(&tx, &result);
                     }
 
                     if result.is_ok() {
-                        result = vcs.checkout(&dir, project);
+                        result = vcs.checkout(&dir, project, pb.as_ref(), force);
                         send_result(&tx, &result);
                     }
 
@@ -152,24 +170,42 @@ impl ManifestInstance {
                         result = execute_actions(&dir, project);
                         send_result(&tx, &result);
                     }
+
+                    if let Some(pb) = pb {
+                        pb.finish();
+                    }
                 });
             }
             drop(tx);
         });
 
-        // Return the first error code if any
-        for code in rx.iter() {
-            match code {
-                ExitCode::Success => {}
-                _ => return Err(code),
+        // Save error messages if any
+        let errors: Vec<String> = rx
+            .iter()
+            .filter_map(|code| match code {
+                Ok(_) => None,
+                Err(msg) => Some(msg),
+            })
+            .collect();
+
+        // If there is any error, return it
+        if !errors.is_empty() {
+            // Save all error message in string, separated by a new line
+            let mut error_msg = String::new();
+            error_msg.push_str("\n\n");
+
+            for msg in errors.iter() {
+                error_msg.push_str(msg);
+                error_msg.push('\n');
             }
+
+            return Err(error_msg);
         }
 
-        debug!("Sync done");
         Ok(())
     }
 
-    pub fn pin<T>(&self, vcs: &T, parser: &dyn ManifestParser) -> Result<Self, ExitCode>
+    pub fn pin<T>(&self, vcs: &T, parser: &dyn ManifestParser) -> Result<Self, String>
     where
         T: VersionControl,
     {
@@ -199,39 +235,35 @@ impl ManifestInstance {
     }
 }
 
-pub fn generate_default_manifest(destination: &String) -> Result<(), ExitCode> {
+pub fn generate_default_manifest(destination: &String) -> Result<(), String> {
     let file = File::create(destination);
 
     match file {
         Ok(_) => {}
-        Err(_) => return Err(ExitCode::GenerationFailed),
+        Err(_) => return Err(format!("Failed to create manifest file {destination}")),
     }
 
     let res = file.unwrap().write_all(DEFAULT_MANIFEST_FILE.as_bytes());
     match res {
         Ok(_) => {}
-        Err(_) => return Err(ExitCode::GenerationFailed),
+        Err(_) => return Err(format!("Failed to write to manifest file {destination}")),
     }
     Ok(())
 }
 
-fn read_manifest(filename: Option<&String>) -> Result<(String, String), ExitCode> {
+fn read_manifest(filename: Option<&String>) -> Result<(String, String), String> {
     let default = MANIFEST_INPUT_DEFAULT.to_string();
     let filename = filename.unwrap_or(&default);
     let file_path = Path::new(&filename);
 
     if !file_path.exists() {
-        error!("{filename} does not exist");
-        return Err(ExitCode::NoManifest);
+        return Err(format!("{filename} does not exist"));
     }
 
-    let file = std::fs::read_to_string(file_path);
+    let file = fs::read_to_string(file_path);
     match file {
         Ok(file) => Ok((filename.clone(), file)),
-        Err(e) => {
-            error!("{e}");
-            Err(ExitCode::NoManifest)
-        }
+        Err(e) => Err(e.to_string()),
     }
 }
 
@@ -250,19 +282,15 @@ fn is_ok_to_clone(manifest_dir: &String, path: &String) -> bool {
     }
 }
 
-fn send_result(sender: &Arc<Mutex<Sender<ExitCode>>>, code: &Result<(), ExitCode>) {
-    let exit_code = match code {
-        Ok(_) => ExitCode::Success,
-        Err(e) => e.clone(),
-    };
+fn send_result(sender: &Arc<Mutex<Sender<Result<(), String>>>>, code: &Result<(), String>) {
     sender
         .lock()
         .expect("Failed to lock mutex.")
-        .send(exit_code)
+        .send(code.clone())
         .expect("Failed to send result through channel.");
 }
 
-fn execute_actions(manifest_dir: &String, project: &Project) -> Result<(), ExitCode> {
+fn execute_actions(manifest_dir: &String, project: &Project) -> Result<(), String> {
     for action in project.get_actions() {
         match action {
             ProjectAction::FileAction(ProjectFileAction::LinkFile(src, dest)) => {
@@ -276,12 +304,11 @@ fn execute_actions(manifest_dir: &String, project: &Project) -> Result<(), ExitC
                 match res {
                     Ok(_) => {}
                     Err(_) => {
-                        error!(
+                        return Err(format!(
                             "Failed to create symlink {} to {}",
                             src.to_string_lossy(),
                             dest.to_string_lossy()
-                        );
-                        return Err(ExitCode::CreateSymlinkFailed);
+                        ));
                     }
                 }
             }
@@ -289,16 +316,44 @@ fn execute_actions(manifest_dir: &String, project: &Project) -> Result<(), ExitC
                 let src = Path::new(manifest_dir).join(project.get_path()).join(src);
                 let dest = Path::new(manifest_dir).join(dest);
                 prepare_file_destination(&dest)?;
-                let res = std::fs::copy(&src, &dest);
+                let res = fs::copy(&src, &dest);
                 match res {
                     Ok(_) => {}
                     Err(_) => {
-                        error!(
+                        return Err(format!(
                             "Failed to copy file {} to {}",
                             src.to_string_lossy(),
                             dest.to_string_lossy()
-                        );
-                        return Err(ExitCode::CopyFailed);
+                        ));
+                    }
+                }
+            }
+            ProjectAction::FileAction(ProjectFileAction::CopyDir(src, dest)) => {
+                let src = Path::new(manifest_dir).join(project.get_path()).join(src);
+                let dest = Path::new(manifest_dir).join(dest);
+                prepare_file_destination(&dest)?;
+                let res = copy_directory(&src, &dest);
+                match res {
+                    Ok(_) => {}
+                    Err(_) => {
+                        return Err(format!(
+                            "Failed to copy directory {} to {}",
+                            src.to_string_lossy(),
+                            dest.to_string_lossy()
+                        ));
+                    }
+                }
+            }
+            ProjectAction::DeleteProject => {
+                let dest = Path::new(manifest_dir).join(project.get_path());
+                let res = fs::remove_dir_all(&dest);
+                match res {
+                    Ok(_) => {}
+                    Err(_) => {
+                        return Err(format!(
+                            "Failed to remove directory {}",
+                            dest.to_string_lossy()
+                        ));
                     }
                 }
             }
@@ -307,12 +362,11 @@ fn execute_actions(manifest_dir: &String, project: &Project) -> Result<(), ExitC
     Ok(())
 }
 
-pub fn save_file(filename: &String, content: &String) -> Result<(), ExitCode> {
+pub fn save_file(filename: &String, content: &String) -> Result<(), String> {
     let mut file = match File::create(filename) {
         Ok(value) => value,
         Err(_) => {
-            error!("Failed to create file {filename}");
-            return Err(ExitCode::SaveFailed);
+            return Err(format!("Failed to create file {filename}"));
         }
     };
 
@@ -320,41 +374,36 @@ pub fn save_file(filename: &String, content: &String) -> Result<(), ExitCode> {
     match res {
         Ok(_) => {}
         Err(_) => {
-            error!("Failed to write to file {filename}");
-            return Err(ExitCode::SaveFailed);
+            return Err(format!("Failed to write to file {filename}"));
         }
     }
     Ok(())
 }
 
-pub fn assert_dependencies() -> Result<(), ExitCode> {
+pub fn assert_dependencies() -> Result<(), String> {
     let output = Command::new("git").arg("--version").output();
 
     match output {
         Ok(_) => Ok(()),
-        Err(_) => {
-            error!("Git is not installed");
-            Err(ExitCode::DependencyMissing)
-        }
+        Err(_) => Err("Git is not installed".to_string()),
     }
 }
 
-fn prepare_file_destination(dest: &PathBuf) -> Result<(), ExitCode> {
+fn prepare_file_destination(dest: &PathBuf) -> Result<(), String> {
     // Delete destination if it exists
     if dest.exists() {
-        let res = std::fs::remove_file(dest);
+        let res = fs::remove_file(dest);
         match res {
             Ok(_) => {}
             Err(_) => {
-                error!("Failed to remove file {}", dest.to_string_lossy());
-                return Err(ExitCode::RemoveFailed);
+                return Err(format!("Failed to remove file {}", dest.to_string_lossy()));
             }
         }
     }
 
     // Create folder is needed
     let parent = dest.parent().unwrap_or("./".as_ref());
-    let _ = std::fs::create_dir_all(parent);
+    let _ = fs::create_dir_all(parent);
 
     Ok(())
 }
@@ -367,4 +416,28 @@ fn get_relative_path(path: &Path, base: &Path) -> PathBuf {
 
     let path = pathdiff::diff_paths(path_dir, base_dir).unwrap_or("./".into());
     path.join(path_filename)
+}
+
+fn copy_directory(src: &Path, dest: &Path) -> std::io::Result<()> {
+    fs::create_dir_all(dest)?;
+
+    let mut queue = VecDeque::new();
+    queue.push_back(src.to_path_buf());
+
+    while let Some(src_path) = queue.pop_front() {
+        let dest_path = dest.join(src_path.strip_prefix(src).unwrap());
+
+        if src_path.is_dir() {
+            fs::create_dir_all(&dest_path)?;
+
+            for entry in fs::read_dir(src_path)? {
+                let entry = entry?;
+                queue.push_back(entry.path());
+            }
+        } else {
+            fs::copy(&src_path, &dest_path)?;
+        }
+    }
+
+    Ok(())
 }