+++ src/git_version_control.rs
@@ -1,9 +1,14 @@
-use crate::application::{DwlMode, ExitCode, VersionControl};
+use crate::application::{DwlMode, VersionControl};
 use crate::project::Project;
-use log::{debug, error};
-use std::io;
+use indicatif::ProgressBar;
+use log::debug;
+use regex::Regex;
+use std::io::{BufRead, BufReader};
 use std::path::Path;
-use std::process::{Command, Output};
+use std::process::{Command, Stdio};
+
+const DISPLAY_STATUS_SIZE: usize = 3;
+const PROGRESS_REFRESH_RATE_MS: u64 = 10;
 
 pub struct GitVersionControl {}
 
@@ -20,7 +25,14 @@ impl Default for GitVersionControl {
 }
 
 impl VersionControl for GitVersionControl {
-    fn clone(&self, manifest_dir: &str, project: &Project, mode: &DwlMode) -> Result<(), ExitCode> {
+    fn clone(
+        &self,
+        manifest_dir: &str,
+        project: &Project,
+        mode: &DwlMode,
+        pb: Option<&ProgressBar>,
+        lightweight: bool,
+    ) -> Result<(), String> {
         let url = match mode {
             DwlMode::HTTPS => project.get_uri_https(),
             DwlMode::SSH => project.get_uri_ssh(),
@@ -34,59 +46,63 @@ impl VersionControl for GitVersionControl {
             repo_path.display()
         );
 
-        let output = Command::new("git")
-            .current_dir(manifest_dir)
-            .args([
-                "clone",
-                url.as_str(),
-                project.get_path().as_str(),
-                "--quiet",
-            ])
-            .output();
+        let mut command = Command::new("git");
 
-        process_command_output(project, output, ExitCode::CloneFailed)
-    }
-
-    /// Clone a project from a URI to a path with a specific revision. The revision MUST be a branch
-    /// or a tag. Commit ID are not supported.
-    fn clone_lightweight(
-        &self,
-        manifest_dir: &str,
-        project: &Project,
-        mode: &DwlMode,
-    ) -> Result<(), ExitCode> {
-        let url = match mode {
-            DwlMode::HTTPS => project.get_uri_https(),
-            DwlMode::SSH => project.get_uri_ssh(),
-        };
-        let repo_path = Path::new(manifest_dir).join(project.get_path());
-
-        debug!(
-            "Cloning {} into {} @ {}",
-            project.get_name(),
-            repo_path.display(),
-            project.get_revision()
-        );
-
-        let output = Command::new("git")
-            .current_dir(manifest_dir)
-            .args([
+        let args: Vec<&str> = if lightweight {
+            [
                 "clone",
                 "--depth",
                 "1",
                 "--branch",
                 project.get_revision().as_str(),
                 "--single-branch",
-                "--quiet",
+                "--progress",
                 url.as_str(),
                 project.get_path().as_str(),
-            ])
-            .output();
+            ]
+            .to_vec()
+        } else {
+            [
+                "clone",
+                url.as_str(),
+                project.get_path().as_str(),
+                "--progress",
+            ]
+            .to_vec()
+        };
 
-        process_command_output(project, output, ExitCode::CloneFailed)
+        command
+            .current_dir(manifest_dir)
+            .args(args)
+            .stderr(Stdio::piped())
+            .stdout(Stdio::null());
+
+        let display_status = [
+            format!("{} cloning", project.get_path()),
+            format!("{} complete", project.get_path()),
+            format!("{} ERROR", project.get_path()),
+        ];
+
+        match process_command(&mut command, pb, &display_status) {
+            Ok(_) => {}
+            Err(e) => {
+                return Err(format!("[{}] Cloning error: \n{}\n", project.get_path(), e));
+            }
+        }
+
+        if let Some(pb) = pb {
+            pb.finish();
+        }
+        Ok(())
     }
 
-    fn checkout(&self, manifest_dir: &str, project: &Project) -> Result<(), ExitCode> {
+    fn checkout(
+        &self,
+        manifest_dir: &str,
+        project: &Project,
+        pb: Option<&ProgressBar>,
+        force: bool,
+    ) -> Result<(), String> {
         let repo_path = Path::new(manifest_dir).join(project.get_path());
 
         debug!(
@@ -96,32 +112,110 @@ impl VersionControl for GitVersionControl {
             project.get_revision()
         );
 
+        let mut command = Command::new("git");
+        command
+            .current_dir(&repo_path)
+            .args(["fetch", "--prune", "--progress"])
+            .stderr(Stdio::piped())
+            .stdout(Stdio::null());
+
+        let display_status = [
+            format!("{} fetch", project.get_path()),
+            format!("{} complete", project.get_path()),
+            format!("{} ERROR", project.get_path()),
+        ];
+        match process_command(&mut command, pb, &display_status) {
+            Ok(_) => {}
+            Err(e) => {
+                return Err(format!("[{}] fetch error: \n{}\n", project.get_path(), e));
+            }
+        }
+
+        let args = if force {
+            ["checkout", project.get_revision(), "--progress", "--force"].to_vec()
+        } else {
+            ["checkout", project.get_revision(), "--progress"].to_vec()
+        };
+        let mut command = Command::new("git");
+        command
+            .current_dir(&repo_path)
+            .args(args)
+            .stderr(Stdio::piped())
+            .stdout(Stdio::null());
+
+        let display_status = [
+            format!("{} checkout", project.get_path()),
+            format!("{} complete", project.get_path()),
+            format!("{} ERROR", project.get_path()),
+        ];
+        match process_command(&mut command, pb, &display_status) {
+            Ok(_) => {}
+            Err(e) => {
+                return Err(format!(
+                    "[{}] checkout error: \n{}\n",
+                    project.get_path(),
+                    e
+                ));
+            }
+        }
+
+        // Check if repository is dirty
         let output = Command::new("git")
             .current_dir(&repo_path)
-            .args(["fetch", "--prune", "--quiet"])
+            .args(["status", "--porcelain", "--untracked-files=no"])
             .output();
 
-        process_command_output(project, output, ExitCode::CheckoutFailed)?;
-
-        let output = Command::new("git")
-            .current_dir(&repo_path)
-            .args(["checkout", project.get_revision(), "--quiet"])
-            .output();
-
-        process_command_output(project, output, ExitCode::CheckoutFailed)?;
+        match output {
+            Ok(output) => {
+                let message = String::from_utf8_lossy(&output.stdout).to_string();
+                if !message.is_empty() {
+                    if let Some(pb) = pb {
+                        pb.set_message(format!("{} ERROR", project.get_path()));
+                    }
+                    return Err(format!(
+                        "[{}] repository is dirty, please commit or stash your changes",
+                        project.get_path()
+                    ));
+                }
+            }
+            Err(e) => {
+                return Err(format!(
+                    "[{}] failed to check repository status: \n{}\n",
+                    project.get_path(),
+                    e
+                ));
+            }
+        }
 
         if is_branch(manifest_dir, project) {
-            let output = Command::new("git")
+            let mut command = Command::new("git");
+            command
                 .current_dir(&repo_path)
-                .args(["pull", "--ff-only", "--quiet"])
-                .output();
+                .args(["merge", "--ff-only", "--progress"])
+                .stderr(Stdio::piped())
+                .stdout(Stdio::null());
 
-            process_command_output(project, output, ExitCode::CheckoutFailed)?;
+            let display_status = [
+                format!("{} merge", project.get_path()),
+                format!("{} complete", project.get_path()),
+                format!("{} ERROR", project.get_path()),
+            ];
+
+            match process_command(&mut command, pb, &display_status) {
+                Ok(_) => {}
+                Err(e) => {
+                    return Err(format!("[{}] merge error: \n{}\n", project.get_path(), e));
+                }
+            }
+        }
+
+        if let Some(pb) = pb {
+            pb.finish();
         }
         Ok(())
     }
 
-    fn get_commit_id(&self, manifest_dir: &str, project: &Project) -> Result<String, ExitCode> {
+    fn get_commit_id(&self, manifest_dir: &str, project: &Project) -> Result<String, String> {
         let repo_path = Path::new(manifest_dir).join(project.get_path());
 
         debug!(
@@ -142,43 +236,70 @@ impl VersionControl for GitVersionControl {
                     .replace(['\n', '\r'], "");
                 Ok(message)
             }
-            Err(e) => {
-                error!("{e}");
-                Err(ExitCode::GetCommitIdFailed)
-            }
+            Err(e) => Err(format!("ERROR [{path}]: {e}", path = project.get_path())),
         }
     }
 }
 
 unsafe impl Sync for GitVersionControl {}
 
-fn process_command_output(
-    project: &Project,
-    output: io::Result<Output>,
-    code: ExitCode,
-) -> Result<(), ExitCode> {
-    match output {
-        Ok(output) => {
-            let message = String::from_utf8_lossy(&output.stderr).to_string();
-            if is_message_ok(&message) {
-                Ok(())
-            } else {
-                error!("[{}] {}", project.get_name(), message);
-                Err(code)
+fn process_command(
+    command: &mut Command,
+    pb: Option<&ProgressBar>,
+    message: &[String; DISPLAY_STATUS_SIZE],
+) -> Result<(), String> {
+    const START_INDEX: usize = 0;
+    const COMPLETE_INDEX: usize = 1;
+    const ERROR_INDEX: usize = 2;
+
+    let mut child = command.spawn().expect("Failed to launch command");
+
+    let mut stderr_capture = Vec::new();
+
+    let re = Regex::new(r"(\d+)%").unwrap();
+
+    if let Some(pb) = pb {
+        pb.set_message(message[START_INDEX].clone());
+    }
+
+    let stderr = BufReader::new(child.stderr.take().expect("Failed to get stderr"));
+    for line in stderr.lines() {
+        let line = line.expect("Failed to read line");
+        stderr_capture.push(line.clone());
+
+        // If the line contains progress information, update the progress bar
+        if let Some(cap) = re.captures(&line) {
+            if let Some(match_) = cap.get(1) {
+                if let Ok(progress) = match_.as_str().parse::<u64>() {
+                    if let Some(pb) = pb {
+                        pb.set_position(progress);
+                    }
+                }
             }
         }
-        Err(e) => {
-            error!("{e}");
-            Err(code)
-        }
+        // Sleep for a short time to avoid hogging the CPU
+        std::thread::sleep(std::time::Duration::from_millis(PROGRESS_REFRESH_RATE_MS));
     }
-}
 
-fn is_message_ok(message: &str) -> bool {
-    if message.contains("fatal") || message.contains("error") {
-        return false;
+    // Check for error message in stderr_capture
+    let error_message: String = stderr_capture
+        .iter()
+        .filter(|line| line.contains("fatal") || line.contains("error"))
+        .fold(String::new(), |acc, line| acc + line + "\n");
+
+    if !error_message.is_empty() {
+        if let Some(pb) = pb {
+            pb.set_message(message[ERROR_INDEX].clone());
+        }
+        return Err(error_message);
     }
-    true
+
+    child.wait().expect("Failed to wait for child process");
+    if let Some(pb) = pb {
+        pb.set_message(message[COMPLETE_INDEX].clone());
+    }
+
+    Ok(())
 }
 
 fn is_branch(manifest_dir: &str, project: &Project) -> bool {