+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-02-20
+
+### Added
+
+- Progress bar when synchronizing the projects.
+- Lightweight clone.
+- Quiet mode.
+- Force option to overwrite local changes.
+- Copy a directory recursively action.
+- Delete repository action.
+
+### Changed
+
+- When synchronizing the projects, using fetch, checkout, merge instead of fetch, checkout, pull.
+- Improve error messages. Now indicating where the error occurred.
+- Error message are message from git with "error" or "fatal" in the message.
+- Testing if the repository is dirty instead of relying on git error message.
+- Add the copydir and delete_project action to the default manifest.
+- Changed the default manifest default parameter from "gitlab.com" to "hostname.com".
+
 ## [0.2.0] - 2023-08-04
 
 ### Changed